fix(world,api): move AdminNpcGateway to :world.core to satisfy modulith
`:api` is only allowed to depend on the umbrella `:world` module per
ModularityTests; the gateway interface was placed in `:world:environment`
(a submodule) which tripped the modulith violation. Move the public
interface + error types + KillOutcome to `:world.core` (package
`dev.gvart.genesara.world`); keep the impl in `:world:environment`.
