+++ api/src/test/kotlin/dev/gvart/genesara/api/internal/rest/admin/worlds/NpcInstanceControllerTest.kt
@@ -26,9 +26,9 @@ import dev.gvart.genesara.world.Terrain
 import dev.gvart.genesara.world.Vec3
 import dev.gvart.genesara.world.WorldId
 import dev.gvart.genesara.world.WorldQueryGateway
-import dev.gvart.genesara.world.environment.AdminNpcGateway
-import dev.gvart.genesara.world.environment.AdminNpcGatewayError
-import dev.gvart.genesara.world.environment.KillOutcome
+import dev.gvart.genesara.world.AdminNpcGateway
+import dev.gvart.genesara.world.AdminNpcGatewayError
+import dev.gvart.genesara.world.KillOutcome
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.springframework.http.MediaType
